changed the dashboard layout of the goverment site and fix the layout.
